+++ packages/office-ui-fabric-react/src/components/Icon/Icon.base.tsx
