+++ packages/office-ui-fabric-react/src/components/Icon/index.ts
@@ -1,2 +1,2 @@
-export * from './Icon';
+export * from './Icon.base';
 export * from './Icon.types';
