+++ packages/office-ui-fabric-react/src/components/Label/Label.base.tsx
